refactor(decoder): new module for the shared decode core
Extract the stable-prefix stabilizer, token-level thought/answer split,
settle/salvage fallbacks, and the chat StreamDecoder out of src/chat/decode.rs
into a new top-level src/decoder/. Move StepProgressEvent out of metal into the
decoder module so the generator depends on the decode contract rather than the
reverse; decoder no longer depends on metal. chat and server import the shared
symbols from crate::decoder.

Comments in the moved code reworded to positive form, free of em dashes and
prose semicolons.
