+++ src/decoder/mod.rs
@@ -1,0 +1,61 @@
+//! The decoding core shared by every stream consumer. A [`Stabilizer`] turns
+//! raw per-step canvas snapshots (`StepProgressEvent`) into stable stop-cut id
+//! prefixes. A token-level [`ChannelIds::split`] separates thought-channel ids
+//! from answer ids, staying quote-aware so channel markup inside a quoted tool
+//! arg reads as data rather than grammar. [`StreamDecoder`] composes the two
+//! into semantic [`ChatEvent`]s. The serve wire mapper builds on the same
+//! primitives, keeping chat and serve in step on the fragile stable-prefix and
+//! channel-split rules.
+
+mod split;
+mod stabilize;
+mod stream;
+#[cfg(test)]
+mod tests;
+
+pub use stream::StreamDecoder;
+
+pub(crate) use split::{ChannelIds, salvage_answer};
+pub(crate) use stabilize::Stabilizer;
+
+/// Per-denoise-step progress snapshot, the contract between the generator and
+/// every stream decoder. `argmax` is the active canvas slice only, dropping
+/// stale rows past `active_canvas` after shrink-on-retry. Under the MLX-exact
+/// sampler the block's final commit is that slice, so a stable prefix is a
+/// faithful preview.
+pub struct StepProgressEvent<'a> {
+    /// 1-based committed-block index this step belongs to.
+    pub block_idx: usize,
+    /// Total blocks budgeted for this generate (`max_new_tokens / CANVAS`).
+    pub max_blocks: usize,
+    /// 1-based denoise step within the block.
+    pub step_in_block: u32,
+    pub max_steps: usize,
+    /// Active-canvas argmax of this step (`active_canvas` entries).
+    pub argmax: &'a [u32],
+    /// Positions accepted by the entropy-bound rule this step.
+    pub accept_count: u32,
+    /// Mean per-position entropy (nats) this step.
+    pub mean_entropy: f32,
+    /// True once this block's argmax is finalized and past retry. A mid-attempt
+    /// "would-stop" step stays false so stream mappers never commit a discarded
+    /// empty or degenerate canvas, which once spliced a second
+    /// `<|channel>thought` into `reasoning_content`.
+    pub block_done: bool,
+}
+
+/// Token-id to text decoding, abstracted so the decoder is testable without the
+/// real 32 MB tokenizer.
+pub trait TextDecoder {
+    fn decode(&self, ids: &[u32]) -> String;
+    fn decode_append(&self, out: &mut String, ids: &[u32]);
+}
+
+impl TextDecoder for std::sync::Arc<crate::tokenizer::Tokenizer> {
+    fn decode(&self, ids: &[u32]) -> String {
+        (**self).decode(ids)
+    }
+    fn decode_append(&self, out: &mut String, ids: &[u32]) {
+        (**self).decode_append(out, ids);
+    }
+}
